UPSTREAM: crypto: chacha - Fix keystream alignment for chacha_block()
commit 9f480faec58cd6197a007ea1dcac6b7c3daf1139 upstream.

When chacha_block() outputs the keystream block, it uses 'u32' stores
directly.  However, the callers (crypto/chacha_generic.c and
drivers/char/random.c) declare the keystream buffer as a 'u8' array,
which is not guaranteed to have the needed alignment.

Fix it by having both callers declare the keystream as a 'u32' array.
For now this is preferable to switching over to the unaligned access
macros because chacha_block() is only being used in cases where we can
easily control the alignment (stack buffers).

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
[cyberknight777: adapt to msm-4.14]
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Abhishek001konni <Abhishek001konni@gmail.com>